HACK: input: touchscreen: goodix: fix driver autoprobing
Because with OF we can pass more specific compatible flags
(such as gt9271) instead of generic Goodix-TS, when i2c-core's
of_i2c_register_devices() tries to request_module(), it'll request it
with a non-existent specific module alias.

In order to have this driver autoprobing again, we just need to add
missing MODULE_ALIAS() entries to goodix driver.

A similar hack patch was made for the edt-ft5x06 driver.

This patch is marked as a "HACK" since this is not acceptable upstream.
Currently the proper fix is still being debated upstream.
https://patchwork.ozlabs.org/patch/618496/
https://patchwork.ozlabs.org/patch/618502/
https://patchwork.ozlabs.org/patch/618494/
https://patchwork.ozlabs.org/patch/618495/
https://patchwork.ozlabs.org/patch/618491/
https://patchwork.ozlabs.org/patch/618493/
https://patchwork.ozlabs.org/patch/618492/
https://patchwork.ozlabs.org/patch/618490/

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>