Input: goodix - put reset and irq into default state during probe
The Goodix touchscreen controller requires a specific pin sequencing to be
done when the device is out of reset. Don't assume the initial state of the
pins at probe. Explicitly put the touchscreen controller into a reset state
and set the interrupt pin to low.

Also don't allow reset pin to float by configuring it as an input. This
avoids assumptions based on board's hardware.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>