mmc: host: omap_hsmmc: disable DEB interrupt during tuning
When clock tuning is in progress for high speed operation
it's expected that there may be data related failures
such as DEB because clk-data lines go out of sync
so do not act upon such errors.
A simple way to avoid acting upon DEB error is to just
disable the event altogether.

This finding will be reflected in future revisions of TRM.

Reviewed-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Ravikumar Kattekola <rk@ti.com>
[nsekhar@ti.com: drop stray empty line]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>