Documentation: dt: keystone: Add new binding for K2G ICE
The 66AK2Gx Industrial Communication Engine (ICE) board is a
board based on TI's 66AK2Gx SOC which has a single core 600 MHz
ARM Cortex-A15 processor. This board is a development platform
for the Industrial market with:
 - 512MB of DDR3L
 - One 1Gbps Ethernet (CPSW)
 - Four 10/100 Industrial Ethernet (PRU-ICSS)
 - uSD
 - QSPI Flash
 - PCIe endpoint
 - USB JTAG and serial console
 - OLED monochrome display
 - Rotary switch
 - Industrial LEDs and PRU expansion connector

Detailed documentation for 66AK2Gx SoC can be found at:

   http://www.ti.com/product/66ak2g02/technicaldocuments

This patch adds a new compatible for the 66AK2Gx ICE board.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
[nsekhar@ti.com: patch description updates
		 call the board ICE, not ICE EVM]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>