ARM: dts: k2g: Add support for K2G ICE
The K2G Industrial Communication Engine (ICE) board is a
board based on TI's 66AK2Gx SOC which has a single core 600 MHz
ARM Cortex-A15 processor. This board is a development platform
for the Industrial market with:
 - 512MB of DDR3L
 - One 1Gbps Ethernet (CPSW)
 - Four 10/100 Industrial Ethernet (PRU-ICSS)
 - uSD
 - QSPI Flash
 - PCIe endpoint
 - USB JTAG and serial console
 - OLED monochrome display
 - Rotary switch
 - Industrial LEDs and PRU expansion connector

Detailed documentation for 66AK2Gx SoC can be found at:

   http://www.ti.com/product/66ak2g02/technicaldocuments

This patch adds a basic device tree file for K2G ICE board.

Only UART and MMC are enabled, and boot to command prompt is
tested.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
[nsekhar@ti.com: patch description updates
		 add uSD card detect
		 dont call the board ICE EVM, just ICE]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
[t-kristo@ti.com: dropped dysfunctional MMC support]
Signed-off-by: Tero Kristo <t-kristo@ti.com>