ARM: dts: keystone-k2g-ice: Add pinmuxes for PRUSS Ethernet
Add the pinmux configurations for enabling the PRUSS Ethernet usecase
on the K2G-ICE board. The PRUSS Ethernet pins have to be added to the
PRUSS nodes so that the PRUSS PHYs can be bootstrapped properly when
the MDIO node probes for the PHYs.

The PR1_PRU1_GPO1 pin (PADCONFIG_202) cannot be setup currently in
kernel due a possible contention with BOOT_CFG RSTMUX8 register
locking. This is still being debugged.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
[nsekhar@ti.com: commit message updates, remove commented pinmux entry]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>