media: ti-vpe: vip: Fix Raw 8 bit memory corruption issue
After adding support for Raw 8 bit mode capture it was found that
VIP and VPDMA make assumption about the incoming data.

When using a Raw 8 Bit format we are faking a YUV422 16bit format
to have the VPDMA perform the needed byte swap we need to adjust
the pixel width accordingly otherwise the parser will attempt to
collect more pixels then available and the VPDMA transfer will
exceed the allocated frame buffer.

Signed-off-by: Benoit Parrot <bparrot@ti.com>