ti_config_fragments: Add ti_config_fragments and basic configs
This is a port of the ti_config_fragments from the TI Linux 4.4
kernel.

Add the defconfig_builder script to create the defconfigs specific
for the TI SDKs.  This script depends on the map file per the README
file located int he ti_config_fragments directory.

The SoC specific .cfg files are being added for am3, am4, dra7,
k2g, keystone and omap5

The system specific .cfg files are being added as well for LPAE,
multi_v7_prune and systemd

Signed-off-by: Dan Murphy <dmurphy@ti.com>