soc: ti: pm33xx: Add prepare_rtc_resume to RTC mode abort path
Currently during an RTC+DDR cycle the RTC hwmod gets enabled by a call
to prepare_rtc_suspend and then disabled during resume by a call to
prepare_rtc_resume to leave it in the state expected by the kernel. If
for some reason an RTC+DDR cycle fails the regular suspend abort path is
executed and this is missing the call to prepare_rtc_resume which causes
omap_hwmod to complain about attempting to enable the hwmod from the
wrong state.

Move the call to prepare_rtc_resume to so it is called during RTC+DDR
resume path for both success and failure.

Signed-off-by: Dave Gerlach <d-gerlach@ti.com>