TEMP: ARM: dts: keystone-k2g-evm: Reduce and relocate the DDR for MPM usecases
The DDR carveout reserved for the Keystone Multi Proc Manager (MPM)
usecases has been reduced from the current 512 MB to 8 MB on the
Keystone K2G EVM board. The start address of this carveout has also
been relocated to align with the address used on the K2G ICE board.
This allows same firmwares to be run on both the K2G platforms.
This reductions allow better utilization of the DDR memory between
different MPM-based usecases.

Signed-off-by: Suman Anna <s-anna@ti.com>