soc: ti: enhancement to support qmss in NSS
NSS available on K2G SoC has a cut down version of QMSS
called qmss lite. It doesn't have status and explicit push
register space as in QMSS available on other K2 SoCs. So define
reg indices specific to QMSS Lite. This patch introduces
"keystone-navigator-qmss-l" compatibility to identify QMSS Lite
and to customize the dts handling code. Per spec, descriptors with
index less than or equal to regions0_size is in region 0 in the
case of QMSS where as for QMSS Lite, descriptors with index less
than regions0_size is in region 0. So update the size accordingly
in the regions0_size bits of the linking ram size 0 register.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>