soc: ti: provide APIs to support driver probe deferral
The driver needs to take care of probe deferral. So any driver
that depends on knav dma and qmss drivers, for example netcp
network driver, needs to defer probe until knav devices are
ready to service. To do this add an API to query the device
ready status from the knav dma and qmss devices.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>