ARM: OMAP2+: omap_device: Fix HWMOD_NEEDS_REIDLE check to actually test
Currently there is a check if an omap_hwmod has the HWMOD_NEEDS_REIDLE
flag set in the omap_device core that improperly uses bitwise OR to mask
out the bit rather than bitwise AND, meaning the test will always show
the flag is set whether it is or not. Although this is not causing any
issues today this should be fixed to avoid potential pain in the future.

Fixes: f92a644f7b39 ("ARM: OMAP2+: omap_hwmod: Introduce HWMOD_NEEDS_REIDLE")
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>