ARM: dts: k2l: Add DT node for OSR
The OSR (Onchip Static RAM) available in K2L SoC is used by QMSS
as an external linkram. However this is a standalone RAM that can
be used for non QMSS usage as well as from DSP side. So add a
sram device node for the same and add the osr clock to the node.
So remove the clock node from netcp node.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Acked-by: Tero Kristo <t-kristo@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>