net: netcp: add packet accelerator (PA2) netcp module
PA2 hardware available on K2E/L SoCs does function similar to
PA on K2HK, but re-designed and requires a separate hw module
driver. This patch adds a pa2 netcp module to support this.
This has its own message formats to communicate with the PA
PDSP firmware. However there are common netcp module functions
that are re-used from the pa core module. Also common functions
are re-used through the pa_core_ops struct. This driver is also
responsible for loading and running PDSPs with required firmware.

The patch also updates the documentation to reflect the PA2
module support. This is a port of the driver from older kernel
baseline that only supports tx checksum hw acceleration. rx checksum
acceleration support will be added in the future as needed.
To know more about the driver, please check the documentation
netcp-pa.txt, updated as part of this patch

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Jack Manbeck <j-manbeck2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>