usb: dwc3: gadget: usb: dwc3: Fix erratic interrupts and enumeration delay
The existing workaround of forcing DEVSPD to SUPER_SPEED
for HIGH_SPEED ports is causing another side effect
which causes erratic interrupts and delayed gadget
enumeration of upto 2 seconds on High-Speed ports.

The metastability issue was not reproducible and so the
alternative workaround for it where we reset the controller
if the issue is found is not upstreamable.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>