ARM: dts: beagle-x15-common: Enable the system mailboxes 3 and 4
Enable the System Mailboxes 3 and 4 and the corresponding child
sub-mailbox nodes on all the AM57xx BeagleBoard-X15 boards, in
order to communicate with the respective PRU remote cores in each
of the PRUSS processor subsystem from the host. It is assumed that
the PRUSS crossbar is setup properly on the PRU firmware side to
receive the mailbox interrupts on one of its input events.

NOTE:
These mailboxes provide an alternate communication path between
the MPU and PRU. The current default approach is to use the PRU
System Events.

Signed-off-by: Suman Anna <s-anna@ti.com>