remoteproc: Reduce asynchronous request_firmware to auto-boot only
[ Upstream commit 7a20c64ddb3deeb08bbe1ca8e9bcafd3241a5e0e ]

The rproc_add_virtio_devices() requests firmware asynchronously and
triggers boot if the auto_boot flag is set. However, this
asynchronous call seems to be redundant for non auto-boot scenario
since the rproc_boot() would call request_firmware() anyways. Move
the auto_boot check to rproc_add() so that a redundant call to
_request_firmware can be avoided for non auto-boot case.

Signed-off-by: Sarangdhar Joshi <spjoshi@codeaurora.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
[s-anna@ti.com: cherry-pick commit '7a20c64ddb3d' from v4.11]
Signed-off-by: Suman Anna <s-anna@ti.com>