i2c: davinci: Add PM Runtime Support
This patch adds PM runtime support to the I2C Davinci driver.

This was verified on K2G EVM and K2G EVM by interacting with the on board
I2C EEPROM. For the above boards along with OMAPL128 EVM additional
instances of I2C were enabled and disabled to verify clock count usage
properly increased and decreased. For K2G EVM an additional debug patch
was added to the TI SCI clock driver to confirm that the clock driver is
properly being triggered.

Reviewed-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>