ARM: dts: AM33xx: Add the PRUSS DT nodes
Add the DT nodes for the PRU-ICSS on AM33xx family of SoCs.
The AM33xx SoCs contain a single PRU-ICSS instance and is
represented by the pruss-soc-bus node and a child PRUSS node.
PRU-ICSS is not supported on AM3352 SoC though in the AM33xx
family, so the nodes are added in disabled state to the common
am33xx dtsi file. They should be enabled only for boards that
use a SoC containing PRU-ICSS in the derivative board files.

The DT nodes use all standard properties. The regs property
defines the addresses for various modules within the PRUSS,
including all the Data RAMs, the Instruction RAM per core,
the Industrial Ethernet Peripheral (IEP), the Real Time Media
Independent Interface controller (MII_RT), CFG, Debug and
Control sub-modules.

The two PRU cores within the PRUSS and the interrupt controller
are represented as child nodes (so platform devices themselves)
of the PRUSS parent node. The default support for virtio/rpmsg
communication between MPU and a PRU core is provided through
PRU system events instead of OMAP mailboxes. This is the
preferred approach as it eliminates an external peripheral
access from the PRU-side. The difference from MPU would be
minimal in using one versus the other.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Keerthy <j-keerthy@ti.com>