ARM: dts: am572x-idk: Add watchdog timers for IPU1 and DSP1
The watchdog timer information has been added to the IPU1 and DSP1
remote processor device nodes on the AM572x IDK board. This data
is identical to the timers used on the AM57xx EVM board to maintain
firmware compatibility between both the SoC and board variants.
The following timers are used as the watchdog timers,
        DSP1 : GPT10
        IPU1 : GPT7 & GPT8 (one for each Cortex-M4 core)

The MPU-side drivers will use this data to initialize the watchdog
timers, and listen for any watchdog triggers. The BIOS-side code on
IPU1 and DSP1 needs to configure/refresh this timer properly to not
throw a watchdog error.

These timers can be changed or removed as per the system integration
needs, alongside appropriate equivalent changes on the firmware side.

Signed-off-by: Suman Anna <s-anna@ti.com>