ARM: dts: am57xx-idk-common: Add PRU system events for virtio
Two PRU system events "vring" and "kick" have been added to each
PRU node in each of the PRU-ICSS1 and PRU-ICSS2 remote processor
subsystems to enable the virtio/rpmsg communication between MPU
and that PRU core. The additions are done in the am57xx-idk-common
dtsi file, and so are inherited by all the AM57xx IDK boards.

The PRU system events is the preferred approach over using OMAP
mailboxes, as it eliminates an external peripheral access from
the PRU-side, and keeps the interrupt generation internal to the
PRUSS. The PRUSS Crossbar also need not be programmed (leverages
the default programming) unlike when using mailboxes. The difference
from MPU would be minimal in using one versus the other.

Mailboxes can still be used if desired. Either approach would require
that an appropriate firmware image is loaded/booted on the PRU.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>