ti_config_fragments: rpmsg: Enable OMAP remoteproc support
Add support to build the OMAP remoteproc driver and the dependent
drivers - OMAP IOMMU and OMAP Mailbox drivers, required for remote
processor messaging with the IPU and DSP remoteprocs on various
OMAP4+ SoCs.

OMAP Mailbox and OMAP IOMMU are chosen to be built-in, while the
OMAP remoteproc driver is chosen to be built as a module. The
Watchdog support configuration and IOMMU debugfs configuration
options are also enabled by default.

Signed-off-by: Suman Anna <s-anna@ti.com>