soc: ti: wkup_m3_ipc: Fix enable_late_halt sysfs entry
In order to fix a race condition that was possible after the call to
rproc_boot the population of the static m3_ipc_state pointer was moved
until after the first interrupt was received from the wkup_m3. However,
the sysfs entry enable_late_halt was still being created during probe
with this static variable, leading to a NULL pointer being used for the
option variable which caused a kernel dump if the sysfs entry was
accessed or written to.

Fix this by making wkup_m3_ipc_dbg_init use the m3_ipc reference and not
calling until the wkup_m3 is actually available.

Fixes: 6832ad8c9c52 ("soc: ti: wkup_m3_ipc: Fix race condition with rproc_boot")
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>