TEMP: remoteproc/pru: add pru_rproc_set_ctable() function
Some firmware e.g. pru_ethernet expects OS to configure the
CTABLE entries. In particular C28 and C30 that point to
Shared RAM and System SRAM (OCMC) areas.

Provide a way for users to do that through a new API,
pru_rproc_set_ctable(). The API returns 0 on success and
a negative value on error.

NOTE:
This probably should be restricted to the PRU firmwares, as
typically they will be re-programming these when dealing with
a certain block of memory. For one-time configuration, the
client driver and firmware can probably use a fixed address
location in shared memory.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
[s-anna@ti.com: add the NOTE: on patch description, minor cleanups]
Signed-off-by: Suman Anna <s-anna@ti.com>