TEMP: soc: ti: add the keystone_dsp_mem driver
A very simple keystone_dsp_mem driver has been added for TI's
Keystone 2 family of SoCs. This driver provides a user-space
mapping interface for some on-chip Multicore Shared Memory (MSM)
SRAM Memory regions and some portions of the platform board's
DDR memory. This is done to enable the Multi Proc Manager (MPM)
based loader for loading different firmware images into both
DDR and on-chip SRAM in userspace for the various C66 DSP
co-processors on the SoC.

The different MSM RAM regions to be exposed to userspace through
this driver need to be defined as 'reserved' child nodes under
the parent MSM RAM mmio-sram node with a specific compatible
"ti,keystone-dsp-msm-ram" property. Each of the DDR regions to
be exposed are defined using reserved-memory nodes and linked
to the driver through an appropriate 'reg' entry value in the
dspmem node. Multiple discrete regions of either SRAM and/or DDR
can be exposed to userspace by defining similar DTS nodes.

The mapping interfaces are provided through a miscdevice,
expected to have a device name (currently /dev/dspmem, derived
from DT node name) matching the usage within MPM. The mmap logic
itself is based on a mechanism used within the UIO framework.

TODO:
Evaluate the need for a 'dspmem' node as it is not a traditional
hardware device.

Signed-off-by: Suman Anna <s-anna@ti.com>