Change USB class for RNDIS
This is a hack to change the USB class/subclass/protocol for the RNDIS
USB gadget function. According to [1], this is "RNDIS over Ethernet".
This configuration is recognized by rndiscmp.inf in Windows so that the
correct driver loads without the need for a custom signed .inf.
Furthermore, it is ignored on macOS and Linux (so we can use CDC/ECM
instead).

[1]: http://www.usb.org/developers/defined_class/#BaseClassEFh