net/rpmsg: fixup rpmsg_sock_getname() return semantics
The function rpmsg_sock_getname() implements the rpmsg proto
socket's .getname() ops and is used for returning the local
or peer address of a socket. Fix up the function to either
return -ENOTCONN or -EINVAL as per the getsockname() and
getpeername() function semantics. Locking has also been
added in preparation for the error recovery support.

Signed-off-by: Suman Anna <s-anna@ti.com>